Separate startup code into different files
