+++ src/utils/databaseConnection.ts
@@ -1,0 +1,7 @@
+import { MongoClient } from "mongodb";
+
+import "dotenv/config";
+
+import mongodbUri from "./mongodbUri";
+
+export default new MongoClient(mongodbUri, {});
